+++ src/base/globals/nclss/mset.rs
@@ -11,7 +11,7 @@ pub(super) fn mkcls(base: Rc<Class>) -> Rc<Class> {
     let methods = vec![
         NativeFunction::new("len", &["self"], None, |globals, args, _kwargs| {
             let set = Eval::expect_mutable_set(globals, &args[0])?;
-            Ok((set.borrow().len() as i64).into())
+            Ok((set.borrow().len()).into())
         }),
         NativeFunction::new("has", &["self", "key"], None, |globals, args, _kwargs| {
             let set = Eval::expect_mutable_set(globals, &args[0])?;
